requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.2
phonenumbers==8.13.55
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
smsaero==1.1.0
smsaero_api==3.0.0
smsaero_api_async==3.0.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3